requirements.txt
streamlit
pandas
requests
python-dotenv
googletrans==4.0.0-rc1
